backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0.post1
python-multipart==0.0.9
pydantic==2.6.1
librosa==0.10.1
supabase==2.3.4
python-dotenv==1.0.1
